# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "h12 API - V1",
    "description": "“Like for post” entity was added\n\nAn authorized user can like the post",
    "version": "h12"
  },
  "paths": {
...
    "/hometask_12/api/blogs/{blogId}/posts": {
      "get": {
        "tags": ["Blogs"],
        "summary": "Returns all posts for specified blog",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          },
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              }
            }
          },
          "404": {
            "description": "If specificied blog is not exists"
          }
        }
      },
      "post": {
        "tags": ["Blogs"],
        "summary": "Create new post for specific blog",
        "parameters": [
          {
            "name": "blogId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Data for constructing new post entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.BlogPostInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created post",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              }
            }
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "If specified blog doesn't exists"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
...
    "/hometask_12/api/posts/{postId}/like-status": {
      "put": {
        "tags": ["Posts"],
        "summary": "Make like/unlike/dislike/undislike operation",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Like model for make like/dislike/reset operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h11.LikeInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h11.LikeInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h11.LikeInputModel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "If post with specified postId doesn't exists"
          }
        },
        "security": [
          {
            "bearer": ["Policy:", "Roles:", "AuthenticationSchemes:"]
          }
        ]
      }
    },
    "/hometask_12/api/posts/{postId}/comments": {
      "get": {
        "tags": ["Posts"],
        "summary": "Returns comments for specified post",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h11.CommentViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h11.CommentViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h11.CommentViewModel>"
                }
              }
            }
          },
          "404": {
            "description": "If post for passed postId doesn't exist"
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create new comment",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Data for constructing new post entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h06.db.CommentInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h06.db.CommentInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h06.db.CommentInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created post",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h11.CommentViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h11.CommentViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h11.CommentViewModel"
                }
              }
            }
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "If post with specified postId doesn't exists"
          }
        },
        "security": [
          {
            "bearer": ["Policy:", "Roles:", "AuthenticationSchemes:"]
          }
        ]
      }
    },
    "/hometask_12/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Returns all posts",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "pageNumber is number of portions that should be returned",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "pageSize is portions size that should be returned",
            "schema": {
              "maximum": 20,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "description": "Default value: desc",
            "schema": {
              "$ref": "#/components/schemas/h04.SortDirections"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paginator<h12.PostViewModel>"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create new post",
        "requestBody": {
          "description": "Data for constructing new post entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the newly created post",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              }
            }
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
    "/hometask_12/api/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Return post by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of existing post",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/h12.PostViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": ["Posts"],
        "summary": "Update existing post by id with InputModel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Data for updating",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/h02.db.PostInputModel"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "If the inputModel has incorrect values",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete post specified by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Post id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "basic": ["Policy:", "Roles:", "AuthenticationSchemes:basic"]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIErrorResult": {
        "type": "object",
        "properties": {
          "errorsMessages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message with error explanation for certain field",
            "nullable": true
          },
          "field": {
            "type": "string",
            "description": "What field/property of input model has error",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Paginator<h11.CommentViewModel>": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "pagesCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h11.CommentViewModel"
            }
          }
        },
        "additionalProperties": false
      },
      "Paginator<h12.PostViewModel>": {
        "required": ["items"],
        "type": "object",
        "properties": {
          "pagesCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h12.PostViewModel"
            }
          }
        },
        "additionalProperties": false
      },
...
      "h02.db.BlogPostInputModel": {
        "required": ["content", "shortDescription", "title"],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 30,
            "type": "string"
          },
          "shortDescription": {
            "maxLength": 100,
            "type": "string"
          },
          "content": {
            "maxLength": 1000,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "h02.db.PostInputModel": {
        "required": ["blogId", "content", "shortDescription", "title"],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 30,
            "type": "string"
          },
          "shortDescription": {
            "maxLength": 100,
            "type": "string"
          },
          "content": {
            "maxLength": 1000,
            "type": "string"
          },
          "blogId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "h04.SortDirections": {
        "enum": ["asc", "desc"],
        "type": "string"
      },
...
      "h06.db.CommentInputModel": {
        "required": ["content"],
        "type": "object",
        "properties": {
          "content": {
            "maxLength": 300,
            "minLength": 20,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "h06.db.CommentatorInfo": {
        "required": ["userId", "userLogin"],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "userLogin": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "h11.CommentViewModel": {
        "required": ["commentatorInfo", "content", "createdAt", "id"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "commentatorInfo": {
            "$ref": "#/components/schemas/h06.db.CommentatorInfo"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "likesInfo": {
            "$ref": "#/components/schemas/h11.LikesInfoViewModel"
          }
        },
        "additionalProperties": false
      },
      "h11.LikeInputModel": {
        "type": "object",
        "properties": {
          "likeStatus": {
            "$ref": "#/components/schemas/h11.LikeStatus"
          }
        },
        "additionalProperties": false
      },
      "h11.LikeStatus": {
        "enum": ["None", "Like", "Dislike"],
        "type": "string",
        "description": "Send None if you want to unlike\\undislike"
      },
      "h11.LikesInfoViewModel": {
        "type": "object",
        "properties": {
          "likesCount": {
            "type": "integer",
            "description": "Total likes for parent item",
            "format": "int32"
          },
          "dislikesCount": {
            "type": "integer",
            "description": "Total dislikes for parent item",
            "format": "int32"
          },
          "myStatus": {
            "$ref": "#/components/schemas/h11.LikeStatus"
          }
        },
        "additionalProperties": false
      },
      "h12.ExtendedLikesInfoViewModel": {
        "type": "object",
        "properties": {
          "likesCount": {
            "type": "integer",
            "description": "Total likes for parent item",
            "format": "int32"
          },
          "dislikesCount": {
            "type": "integer",
            "description": "Total dislikes for parent item",
            "format": "int32"
          },
          "myStatus": {
            "$ref": "#/components/schemas/h11.LikeStatus"
          },
          "newestLikes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/h12.LikeDetailsViewModel"
            },
            "description": "Last 3 likes (status \"Like\")",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "h12.PostViewModel": {
        "required": [
          "blogId",
          "blogName",
          "content",
          "id",
          "shortDescription",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "blogId": {
            "type": "string"
          },
          "blogName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "extendedLikesInfo": {
            "$ref": "#/components/schemas/h12.ExtendedLikesInfoViewModel"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token **_only_**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "refreshToken": {
        "type": "apiKey",
        "description": "JWT refreshToken inside cookie. Must be correct, and must not expire",
        "name": "refreshToken",
        "in": "cookie"
      }
    }
  }
}